move cms lib module to new lib layout
