+++ projects/storefrontlib/src/cms-components/index.ts
@@ -1,4 +1,5 @@
 export * from './checkout/index';
+export * from './cms-lib.module';
 export * from './content/index';
 export * from './misc/index';
 export * from './myaccount/index';
