--- projects/storefrontlib/src/lib/cms-lib/index.ts
@@ -1,1 +1,0 @@
-export * from './cms-lib.module';
